`SVGHandling` as a separate python module.
